Commit some bzedit changes that have been languishing in my tree for a while Add missing VC7.1 project files for the libs Cleanly compile on VC7.1 Avoid crashing in some situation that I can't remember anymore Correct URLs in About dialog
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@10211 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ bzeditw32/BZEdit32/Model.cpp
@@ -178,7 +178,7 @@ void CModelManager::KillModel ( int iItem )
 
 CBaseModel	*CModelManager::GetModel ( int iItem )
 {
-	if ( (iItem < 0) && (iItem >= aModelList.size()) )
+	if ( (iItem < 0) && (iItem >= (int)aModelList.size()) )
 		return NULL;
 
 	return aModelList[iItem];
@@ -186,7 +186,7 @@ CBaseModel	*CModelManager::GetModel ( int iItem )
 
 void CModelManager::AddDrawItem ( int iItem, bool bTransparent, bool bSelected, int iName, teDrawPriority ePriority )
 {
-	if ( (iItem < 0) && (iItem >= aModelList.size()) )
+	if ( (iItem < 0) && (iItem >= (int) aModelList.size()) )
 		return;
 
 	trDrawItem	rDrawItem;
